Undo xpos

+++ nemo/collections/nlp/modules/common/megatron/xpos_relative_position.py
@@ -61,5 +61,5 @@ class XPOS(nn.Module):
         if downscale:
             scale = 1 / scale
 
-        x = apply_rotary_pos_emb(x, sin.to(x.dtype), cos.to(x.dtype), scale.to(x.dtype))
+        x = apply_rotary_pos_emb(x, sin, cos, scale)
         return x
